rename `operation_tags` module
