+++ fedot/core/repository/operation_tags.py
