+++ fedot/core/repository/operation_types_repo_enum.py
@@ -6,7 +6,7 @@ import json
 from pathlib import Path
 from typing import List, Optional
 
-from fedot.core.repository.operation_tags_n_repo_enums import ALL_TAGS, TagsEnum
+from fedot.core.repository.operation_tags import ALL_TAGS, TagsEnum
 
 from fedot.core.repository.dataset_types import DataTypesEnum
 from fedot.core.repository.json_evaluation import import_enums_from_str, import_strategy_from_str, read_field
