+++ fedot/preprocessing/structure.py
@@ -4,7 +4,7 @@ import networkx as nx
 import numpy as np
 import pandas as pd
 
-from fedot.core.repository.operation_tags_n_repo_enums import TagsEnum, OtherTagsEnum, DataOperationTagsEnum
+from fedot.core.repository.operation_tags import TagsEnum, OtherTagsEnum, DataOperationTagsEnum
 from golem.core.dag.convert import graph_structure_as_nx_graph
 
 DEFAULT_SOURCE_NAME = 'default'
