+++ test/integration/validation/test_table_cv.py
@@ -6,7 +6,7 @@ from fedot.core.composer.composer_builder import ComposerBuilder
 from fedot.core.data.data_split import train_test_data_setup
 from fedot.core.pipelines.pipeline import Pipeline
 from fedot.core.pipelines.pipeline_composer_requirements import PipelineComposerRequirements
-from fedot.core.repository.operation_tags_n_repo_enums import ModelTagsEnum
+from fedot.core.repository.operation_tags import ModelTagsEnum
 from fedot.core.repository.operation_types_repository import OperationTypesRepository
 from fedot.core.repository.metrics_repository import ClassificationMetricsEnum
 from fedot.core.repository.tasks import Task, TaskTypesEnum
